Extract system test base code from LogFileTest
Created a base class for system tests which allows for running the base controller safely from a test. It uses exceptions instead of allowing the program to exit. This allows for catching those inside actual tests.

Initially written for log file test.

Driver: @M4eqP
Navigator: @TobiCode
