+++ cloud.foundry.cli/src/test/java/cloud/foundry/cli/system/SystemExitException.java
@@ -1,0 +1,16 @@
+package cloud.foundry.cli.system;
+
+/**
+ * Thrown by custom security manager whenever System.exit(...) is called.
+ */
+public class SystemExitException extends SecurityException {
+    private final int exitCode;
+
+    public SystemExitException(int exitCode) {
+        this.exitCode = exitCode;
+    }
+
+    public int getExitCode() {
+        return exitCode;
+    }
+}
